+++ migrations/6_configs.up.sql
@@ -1,0 +1,7 @@
+INSERT INTO
+    `config`(`key`, `value`)
+VALUES
+    ('SPOTIFY_API_ID', ''),
+    ('SPOTIFY_API_SECRET', ''),
+    ('LASTFM_API_KEY', ''),
+    ('REGISTRATION_ENABLED', '1');